RealTime/kimiRealTime/requirements.txt
# lib
crc==1.3.0
# basics
Flask==2.2.2
Werkzeug==2.2.2
Jinja2==3.1.2
pyserial==3.5
# animations
APScheduler==3.10.1
numpy==1.23.4
# speech
pygame>2.1.2
# speech on server
requests==2.28.2
# chatGpt demo
openai
python-dotenv==0.21.0
sounddevice==0.4.6
soundfile==0.12.1
# posedetection
websockets==12.0.*
# singing
librosa==0.10.*
scipy==1.11.4
# external button listener
keyboard>0.13
